strangerchatbot/requirements.txt
telepot
psycopg2-binary